fix: bridge jotai
