+++ src/state/index.ts
@@ -3,7 +3,6 @@ import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit';
 import React from 'react';
 import { createDispatchHook, createSelectorHook, createStoreHook } from 'react-redux';
 import { load, save } from 'redux-localstorage-simple';
-import pbridge from './pbridge/reducer';
 import plists from './plists/reducer';
 import pmulticall from './pmulticall/reducer';
 
@@ -12,7 +11,6 @@ export const PANGOLIN_PERSISTED_KEYS: string[] = ['plists'];
 export const pangolinReducers = {
   plists,
   pmulticall,
-  pbridge,
 };
 
 const store = configureStore({
